Query-WIzard-main/requirements.txt
streamlit
mysql-connector-python
googletrans==4.0.0-rc1
speechrecognition
pyaudio
